requirements.txt
pandas==2.1.4
SQLAlchemy==2.0.25
streamlit==1.30.0
python-dotenv==1.0.0
aiosqlite==0.19.0
